Weird issue with libraryroot.custom.css not apply on clean install

+++ backend.py
@@ -177,6 +177,7 @@ def get_json_data():
             global json_data
             json_data = json.load(f)
         print("Loaded JSON data. " + "(" + json_data_filename + ")")
+        f.close()
         return json_data
     except FileNotFoundError:
         print("JSON file " + json_data_filename + " not found.", file=sys.stderr)
@@ -339,23 +340,27 @@ def write_css_settings(settings, settings_values, root_config):
 # Compiles libraryroot.custom.css from /scss directory
 # Adds variables.css
 def compile_css(json_data):
-    sass.compile(dirname=('scss','.'),
-                 output_style='expanded')
-    with open('libraryroot.custom.css', "r", newline='', encoding="UTF-8") as f, \
-         open("libraryroot.custom.temp.css", "w", newline='', encoding="UTF-8") as f1:
-        for line in f:
-            if json_data["CSSVariableString"] in line:
-                #print(line)
-                with open('variables.css', "r", newline='', encoding="UTF-8") as v1:
-                    for variable_line in v1:
-                        f1.write(variable_line)
-                        #print(variable_line)
-                v1.close()
-                f1.write(OS_line_ending())
-            else:
-                f1.write(line)
-    f.close()
-    f1.close()
+    try:
+        sass.compile(dirname=('scss','.'),
+                     output_style='expanded')
+        with open('libraryroot.custom.css', "r", newline='', encoding="UTF-8") as f, \
+             open("libraryroot.custom.temp.css", "w", newline='', encoding="UTF-8") as f1:
+            for line in f:
+                if json_data["CSSVariableString"] in line:
+                    #print(line)
+                    with open('variables.css', "r", newline='', encoding="UTF-8") as v1:
+                        for variable_line in v1:
+                            f1.write(variable_line)
+                            #print(variable_line)
+                    v1.close()
+                    f1.write(OS_line_ending())
+                else:
+                    f1.write(line)
+        f.close()
+        f1.close()
+    except:
+        print("Error compiling SCSS to CSS", file=sys.stderr)
+        print_traceback()
 
     ###
     shutil.move("libraryroot.custom.css", "libraryroot.custom.css.backup")
@@ -705,23 +710,37 @@ def write_js_fixes(fixesdata, special_fixesdata):
         print_traceback()
                     
 def refresh_steam_dir():
-    shutil.copy2("libraryroot.custom.css", library_dir() + "/" + "libraryroot.custom.css")
-    print("File " + "libraryroot.custom.css" + " written to " + library_dir())
-    
-    #shutil.copy2(library_dir() + "/licenses.txt", library_dir() + "/licenses.txt.copy")
-    #os.remove(library_dir() + "/licenses.txt.copy")
-    f = open(library_dir() + "/refresh_dir.txt", "w", newline='', encoding="UTF-8")
-    f.close()
-    os.remove(library_dir() + "/refresh_dir.txt")
+    try:
+        if os.path.isfile(library_dir() + "/" + "libraryroot.custom.css"):# and os.stat(library_dir() + "/" + "libraryroot.custom.css").st_size > 15:
+            if os.path.isfile(library_dir() + "/" + "libraryroot.custom.css.backup"):
+                print("Existing libraryroot.custom.css code detected.")
+                shutil.copy2(library_dir() + "/" + "libraryroot.custom.css", library_dir() + "/" + "libraryroot.custom.css.backup2")
+                print("Backed up steamui/libraryroot.custom.css to steamui/libraryroot.custom.css.backup2")
+            else:
+                shutil.copy2(library_dir() + "/" + "libraryroot.custom.css", library_dir() + "/" + "libraryroot.custom.css.backup")
+                print("backed up steamui/libraryroot.custom.css to steamui/libraryroot.custom.css.backup")
+            shutil.copy2("libraryroot.custom.css", library_dir() + "/" + "libraryroot.custom.css")
+        elif not os.path.isfile(library_dir() + "/" + "libraryroot.custom.css"):
+            shutil.copy2("libraryroot.custom.css", library_dir() + "/" + "libraryroot.custom.css")
+        print("File " + "libraryroot.custom.css" + " written to " + library_dir())
+        
+        #shutil.copy2(library_dir() + "/licenses.txt", library_dir() + "/licenses.txt.copy")
+        #os.remove(library_dir() + "/licenses.txt.copy")
+        f = open(library_dir() + "/refresh_dir.txt", "w", newline='', encoding="UTF-8")
+        f.close()
+        os.remove(library_dir() + "/refresh_dir.txt")
+    except:
+        print("Unable to copy libraryroot.custom.css to Steam directory.", file=sys.stderr)
+        print_traceback()
     
 
 def clean_slate_css():    
     try:
-        f = open(library_dir() + "/libraryroot.empty.css", "w", newline='', encoding="UTF-8")
+        f = open(library_dir() + "/" + "libraryroot.empty.css", "w", newline='', encoding="UTF-8")
         #f.write(OS_line_ending())
         f.close()
-        
-        shutil.move(library_dir() + "/libraryroot.custom.css", library_dir() + "/libraryroot.custom.css.backup")
+        if os.path.isfile(library_dir() + "/" + "libraryroot.custom.css"):
+            shutil.move(library_dir() + "/libraryroot.custom.css", library_dir() + "/libraryroot.custom.css.backup")
         shutil.move(library_dir() + "/libraryroot.empty.css", library_dir() + "/libraryroot.custom.css")
         print("libraryroot.custom.css in Steam directory emptied out, backup at libraryroot.custom.css.backup")
         shutil.copy2("themes/config.css.original", library_dir() + "/config.css")
